+++ lib/src/typings.dart
@@ -1,6 +1,9 @@
 import 'package:flutter/widgets.dart';
 
-import '../nuvigator.dart';
+import 'next/v1/nu_router.dart';
+import 'nu_route_settings.dart';
+import 'nuvigator.dart';
+import 'screen_route.dart';
 
 typedef ScreenRouteBuilder = ScreenRoute<dynamic> Function(
     RouteSettings settings);
@@ -11,3 +14,12 @@ typedef HandleDeepLinkFn = Future<dynamic> Function(INuRouter router, Uri uri,
 typedef ObserverBuilder = NavigatorObserver Function();
 
 typedef WrapperFn = Widget Function(BuildContext context, Widget child);
+
+typedef NuWidgetRouteBuilder = Widget Function(
+    BuildContext context, NuRoute nuRoute, NuRouteSettings<dynamic> settings);
+
+typedef NuRouteParametersParser<A> = A Function(Map<String, dynamic>);
+
+typedef NuInitFunction = Future<bool> Function(BuildContext context);
+
+typedef ParamsParser<T> = T Function(Map<String, dynamic> map);